backend/requirements.txt
fastapi==0.92.0
python-box==7.0.0
pyyaml==6.0
uvicorn==0.20.0
